refactor deps into separate module (no circular imports)
