+++ sample_plugin/api.py
@@ -4,8 +4,11 @@ from contextlib import contextmanager
 from fastapi import APIRouter
 
 from quetz.db_models import PackageVersion
+from quetz.deps import get_db
 
 router = APIRouter()
+
+get_db = contextmanager(get_db)
 
 
 @router.get(
@@ -15,9 +18,6 @@ def get_plugin(channel_name, package_name, version_hash):
 
     version_id, build_string = version_hash.split("-")
 
-    from quetz.main import get_db
-
-    get_db = contextmanager(get_db)
     with get_db() as db:
         package_version = (
             db.query(PackageVersion)
