+++ sample_plugin/main.py
@@ -6,6 +6,9 @@ import db_models
 from api import router
 
 import quetz
+from quetz.deps import get_db
+
+get_db = contextmanager(get_db)
 
 
 @quetz.hookimpl
@@ -16,9 +19,8 @@ def register_router():
 @quetz.hookimpl
 def post_add_package_version(version, condainfo):
     run_exports = json.dumps(condainfo.run_exports)
-    from quetz.main import get_db
 
-    with contextmanager(get_db)() as db:
+    with get_db() as db:
         metadata = db_models.PackageVersionMetadata(
             id=uuid.uuid4().bytes, version_id=version.id, run_exports=run_exports
         )
